+++ metadata.rb
@@ -4,7 +4,7 @@ maintainer_email  'help@sous-chefs.org'
 license           'Apache 2.0'
 description       'Installs and configures haproxy'
 long_description  IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version           '2.0.2'
+version           '3.0.0'
 
 recipe 'haproxy', 'Installs and configures haproxy'
 recipe 'haproxy::app_lb', 'Installs and configures haproxy by searching for nodes of a particular role'
@@ -15,7 +15,8 @@ end
 
 depends           'cpu', '>= 0.2.0'
 depends           'build-essential'
+depends           'poise-service'
 
-source_url 'https://github.com/sous-chefs/haproxy' if respond_to?(:source_url)
-issues_url 'https://github.com/sous-chefs/haproxy/issues' if respond_to?(:issues_url)
-chef_version '>= 11.0' if respond_to?(:chef_version)
+source_url 'https://github.com/sous-chefs/haproxy'
+issues_url 'https://github.com/sous-chefs/haproxy/issues'
+chef_version '>= 12.1'