+++ spec/recipes/install_package_spec.rb
@@ -11,15 +11,6 @@ describe 'haproxy::install_package' do
     expect(chef_run).to create_directory('/etc/haproxy')
   end
 
-  it 'creates the haproxy init.d script' do
-    expect(chef_run).to create_template('/etc/init.d/haproxy').with(
-      source: 'haproxy-init.erb',
-      owner: 'root',
-      group: 'root',
-      mode: '0755'
-    )
-  end
-
   describe 'with version set' do
     let(:given_version) { '1.2.3.4' }
     cached(:chef_run) do